Added ability to monitor effective sample. This led to a change in the way the posterior expectation is computed for trial peak offsets
Added a running value for the posterior expectation, that is updated at each iteration.
Also found a bug: The values of pi and theta were being updated at the evaluation steps, which likely explains the drops I was seeing with some factors being empty. This might mean correction factor was never needed to begin with